requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
Django==2.2
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
Pillow==6.2.1
pkg-resources==0.0.0
pydub==0.23.1
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
